requirements.txt
pydub 
faster-whisper 
argostranslate
yt-dlp
python-dotenv
openai
transformers 
torch 
sentencepiece
datasets